requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
Authlib==1.3.2
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
ccxt==4.4.25
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
distro==1.9.0
fastapi==0.115.4
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.12.1
numpy==2.1.3
openai==1.53.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pip-tools==7.4.1
plotly==5.24.1
propcache==0.2.0
protobuf==5.28.3
pyarrow==18.0.0
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.2
streamlit==1.39.0
stripe==11.2.0
ta==0.11.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.0
wheel==0.44.0
yarl==1.17.1
